+++ packages/@pufflig/ps-nodes-config/src/inputs/index.ts
@@ -1,5 +1,7 @@
+import { MessageInput } from "./message_input/message_input";
 import { templateEditor } from "./template_editor/template_editor";
 
 export const inputNodes = {
   "input/template_editor": templateEditor,
+  "input/message_input": MessageInput,
 };