+++ packages/@pufflig/ps-nodes-config/src/inputs/message_input/message_input.ts
@@ -1,0 +1,26 @@
+import { NodeConfig } from "@pufflig/ps-types";
+
+export const MessageInput: NodeConfig = {
+  name: "Message Input",
+  description: "",
+  tags: ["input", "chat"],
+  parameters: [],
+  outputs: [
+    {
+      id: "message",
+      name: "Message",
+      description: "",
+      type: "message",
+      defaultValue: null,
+    },
+  ],
+  inputs: [
+    {
+      id: "message",
+      name: "Message",
+      description: "",
+      type: "message",
+      defaultValue: null,
+    },
+  ],
+};