+++ packages/@pufflig/ps-nodes-config/src/outputs/display/display_chat.ts
@@ -12,7 +12,9 @@ export const displayChat: NodeConfig = {
       name: "Chat",
       description: "",
       type: "chat",
-      defaultValue: [],
+      defaultValue: {
+        messages: [],
+      },
     },
   ],
 };